Make AddFrameModal components modular
